Grid_Codebook_Modeling/requirements.txt
easydict==1.9
imageio==2.13.5
opencv-python==4.1.2.30
PyWavelets==1.1.1
pyyaml==5.3.1
six==1.15.0
tensorboardX==2.4.1
timm==0.3.2
tokenizers==0.9.4
tqdm
lmdb==1.2.1
lws==1.2
einops==0.3.2
albumentations==0.5
pudb==2019.2
pytorch-lightning==1.5
omegaconf==2.0.0
test-tube>=0.7.5
more-itertools>=8.0.0
transformers==4.2.0
chardet==3.0.4
